+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2023-11-17
+
+### Added
+
+- Add the issue release/white paper page [#210](https://github.com/policy-design-lab/pdl-frontend/issues/210)
+- Add iFrame component to the release/white paper page [#214](https://github.com/policy-design-lab/pdl-frontend/issues/214)
+
 ## [0.11.0] - 2023-11-08
 
 ### Added